Potential fix for pull request finding

+++ src/game/WorldHandlers/GridMap.cpp
@@ -876,7 +876,7 @@ bool GridMap::ExistMap(uint32 mapid, int gx, int gy)
 {
     int len = sWorld.GetDataPath().length() + strlen("maps/%04u%02u%02u.map") + 1;
     char* tmp = new char[len];
-    snprintf(tmp, len, (char*)(sWorld.GetDataPath() + "maps/%04u%02u%02u.map").c_str(), mapid, gx, gy);
+    snprintf(tmp, len, (char*)(sWorld.GetDataPath() + "maps/%04u%02u%02u.map").c_str(), mapid, (unsigned)gx, (unsigned)gy);
 
     FILE* pf = fopen(tmp, "rb");
